Moved files and fixed startup
