+++ android/UbiSolar/UbiSolar/src/main/AndroidManifest.xml
@@ -7,21 +7,20 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".activities.DrawerActivity"
-            android:label="@string/app_name" >
+            android:name=".activities.LoginActivity"
+            android:label="@string/app_name"
+            android:windowSoftInputMode="adjustResize|stateVisible">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <activity
-            android:name="com.sintef_energy.ubisolar.LoginActivity"
-            android:label="@string/title_activity_login"
-            android:windowSoftInputMode="adjustResize|stateVisible" >
+            android:name=".activities.DrawerActivity"
+            android:label="@string/app_name">
         </activity>
     <provider
-            android:name="com.sintef_energy.ubisolar.database.energy.EnergyProvider"
+            android:name=".database.energy.EnergyProvider"
             android:authorities="com.sintef_energy.ubisolar.db.EnergyContract"
             android:exported="true"
             android:grantUriPermissions="true"
@@ -29,5 +28,4 @@
             android:readPermission=".alarm.db.AlarmModel.READ"
             android:writePermission=".alarm.db.AlarmModel.WRITE" />
     </application>
-
 </manifest>
