+++ android/UbiSolar/UbiSolar/src/main/res/layout/activity_login.xml
@@ -1,6 +1,6 @@
 <merge xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.sintef_energy.ubisolar.LoginActivity">
+    tools:context="com.sintef_energy.ubisolar.activities.LoginActivity">
 
     <!-- Login progress -->
     <LinearLayout android:id="@+id/login_status"
